Add save and get API routes
Adds SQLite3 Database queries to store and list sensor values.
SQLite3 is chosen due to its simple implementation in python
and avoid a bunch of dependencies.

Database consist four fields: id, key, time and value. Database
get query is designed with flexibility for fields to be added in
future.

save and get API routes are defined to save the value and key
in the database with FastAPI.

Signed-off-by: Pradyumna Krishna <git@onpy.in>